fix: 滚轮偏移2个数字+颜色显示修复
Bug fixes:
- 滚轮选择器: snap逻辑减去3个padding项偏移,选中值与显示一致
- 颜色统一使用0xFF4CAF50.toLong()格式(-1090713120L)
- 颜色比较改为toInt()比较,兼容两种Long格式
- 新建项目颜色默认值与数据库格式一致

+++ app/src/main/java/com/bulubulu/recordlifeitems/ui/projects/ProjectEditScreen.kt
@@ -163,7 +163,8 @@ fun ProjectEditScreen(
             ) {
                 ProjectColors.forEach { color ->
                     val colorLong = color.value.toLong()
-                    val isSelected = colorLong == selectedColor
+                    // Normalize: compare as Int to handle both -1090713120L and 4283215696L
+                    val isSelected = colorLong.toInt() == selectedColor.toInt()
                     Box(
                         modifier = Modifier
                             .size(40.dp)
